Moving generated code to pkg, updating pkg references
