+++ pkg/lib/go/v1/swagger.go
